PaymentService 구현 그 외 다양한 서비스, 컨트롤러 수정 id 타입 Long으로 통일 그 외 자잘한 수정

+++ baseball_ticket_sale/src/main/java/com/example/baseball_ticket_sale/Dto/CancelDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CancelDto {
-    private int cancelNumber;
+    private Long cancelNumber;
     private String cancelStatus;
     private int refundFee;
     private String refundStatus;
